+ shorts as markers instead of byte

+++ JavaEEGUtils/src/it/hakvoort/eeg/marker/MarkerListener.java
@@ -8,6 +8,6 @@ package it.hakvoort.eeg.marker;
  */
 public interface MarkerListener {
 
-	void receivedMarker(byte marker);
+	void receivedMarker(short marker);
 	
 }
